rename customTypes store to availableCustomTypes store
