+++ packages/slice-machine/src/modules/availableCustomTypes/factory.ts
