+++ packages/slice-machine/src/redux/saga.ts
@@ -1,7 +1,7 @@
 import { fork } from "redux-saga/effects";
 
 import { watchSimulatorSagas } from "@src/modules/simulator";
-import { watchCustomTypesSagas } from "@src/modules/customTypes";
+import { watchAvailableCustomTypesSagas } from "@src/modules/availableCustomTypes";
 import { watchSelectedCustomTypeSagas } from "@src/modules/selectedCustomType";
 import { watchSliceSagas } from "@src/modules/slices";
 import { watchToasterSagas } from "@src/modules/toaster";
@@ -9,7 +9,7 @@ import { watchToasterSagas } from "@src/modules/toaster";
 // Single entry point to start all Sagas at once
 export default function* rootSaga() {
   yield fork(watchSimulatorSagas);
-  yield fork(watchCustomTypesSagas);
+  yield fork(watchAvailableCustomTypesSagas);
   yield fork(watchSliceSagas);
   yield fork(watchToasterSagas);
   yield fork(watchSelectedCustomTypeSagas);
