+++ packages/slice-machine/src/redux/type.ts
@@ -3,7 +3,7 @@ import { LoadingStoreType } from "@src/modules/loading/types";
 import { UserContextStoreType } from "@src/modules/userContext/types";
 import { EnvironmentStoreType } from "@src/modules/environment/types";
 import { SimulatorStoreType } from "@src/modules/simulator/types";
-import { CustomTypesStoreType } from "@src/modules/customTypes/types";
+import { AvailableCustomTypesStoreType } from "@src/modules/availableCustomTypes/types";
 import { SelectedCustomTypeStoreType } from "@src/modules/selectedCustomType/types";
 import { SlicesStoreType } from "@src/modules/slices/types";
 import { RouterState } from "connected-next-router/types";
@@ -14,7 +14,7 @@ export type SliceMachineStoreType = {
   userContext: UserContextStoreType;
   environment: EnvironmentStoreType;
   simulator: SimulatorStoreType;
-  customTypes: CustomTypesStoreType;
+  availableCustomTypes: AvailableCustomTypesStoreType;
   selectedCustomType: SelectedCustomTypeStoreType;
   slices: SlicesStoreType;
   router: RouterState;
